Drop the soft-float and 64-bit division from the input normalization path

+++ Core/Src/FAC_Code/fac_app.c
@@ -341,28 +341,31 @@ void FAC_app_init_all_modules(void) {
 
 /**
  * @brief 		change the range of a variable from one to another
+ * @note		The intermediate product is calculated on 32 bit. Cortex-M0 has no divider at all,
+ * 				so every division is a library routine, and the 64 bit one is far longer than the
+ * 				32 bit one. All the values handled by the firmware (channel counts, speeds, servo
+ * 				positions) stay well below the limit: the widest product in use is about 4*10^6
+ * @IMPORTANT	Keep (x - in_min) * (out_max - out_min) below 2^32 or the result wraps around
  * @retval 		return the value in the new range
  */
 uint32_t map_uint32(uint32_t x, uint32_t in_min, uint32_t in_max,
 		uint32_t out_min, uint32_t out_max) {
 	if (x > in_max)
 		x = in_max;
-	// Cast to uint64_t to avoid overflow during the calculation
-	return (uint32_t) (((uint64_t) (x - in_min) * (out_max - out_min))
-			/ (in_max - in_min) + out_min);
+	return (((x - in_min) * (out_max - out_min)) / (in_max - in_min) + out_min);
 }
 
 /**
  * @brief 		change the range of a variable from one to another
+ * @note		The intermediate product is calculated on 32 bit, see map_uint32 for the reason
+ * @IMPORTANT	Keep (x - in_min) * (out_max - out_min) inside the int32_t range or the result overflows
  * @retval 		return the value in the new range
  */
 int32_t map_int32(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min,
 		int32_t out_max) {
 	if (x > in_max)
 		x = in_max;
-	// Cast to uint64_t to avoid overflow during the calculation
-	return (int32_t) (((int64_t) (x - in_min) * (out_max - out_min))
-			/ (in_max - in_min) + out_min);
+	return (((x - in_min) * (out_max - out_min)) / (in_max - in_min) + out_min);
 }
 
 /**
